+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Enhancements
 - Protection against concurrent bg deploys
-- Add `FlyDeploy.blue_green_cluster_status(otp_app)`
+- Add `FlyDeploy.blue_green_cluster_status()`
 - Fix early timeout errors being display on bluegreen deploys
 
 ## 0.3.3 (2026-03-02)